[JSC] WasmOperations: divide the frame offset as a signed value
The `access` lambdas turn a byte offset into an index with `i / sizeof(V)`. `i`
is an int and sizeof is a size_t, so a negative offset is converted to unsigned
first: WasmToJSCallableFunctionSlot (-8) becomes index 0x1FFFFFFFFFFFFFFF, and
&p[0x1FFFFFFFFFFFFFFF] is pointer arithmetic that overflows. It wraps to p - 8,
so the right slot is read, but it is undefined behaviour.

Under ASan on x86_64 the compiler folds the check it puts before that load into
a read of the constant shadow address ((uint64_t)-8 >> 3) + 0x7fff8000, which
is not mapped: operationWasmToJSExitMarshalArguments faulted in the check
itself, and every wasm test that calls into JS with the JIT off (wasm-no-jit,
lockdown) crashed on the linux-amd64-asan lane, 370 files. The division is done
in int.

+++ Source/JavaScriptCore/wasm/WasmOperations.cpp
@@ -107,7 +107,7 @@ JSC_DEFINE_JIT_OPERATION(operationJSToWasmEntryWrapperBuildFrame, JSToWasmCallee
 
     auto access = [sp, callFrame]<typename V>(auto* arr, int i) -> V* {
         dataLogLnIf(WasmOperationsInternal::verbose, "fp[", (&reinterpret_cast<uint8_t*>(arr)[i / sizeof(uint8_t)] - reinterpret_cast<uint8_t*>(callFrame)), "] sp[", (&reinterpret_cast<uint8_t*>(arr)[i / sizeof(uint8_t)] - reinterpret_cast<uint8_t*>(sp)), "](", RawHex(reinterpret_cast<V*>(arr)[i / sizeof(V)]), ")");
-        return &reinterpret_cast<V*>(arr)[i / sizeof(V)];
+        return &reinterpret_cast<V*>(arr)[i / static_cast<int>(sizeof(V))];
     };
 
     CallInformation wasmFrameConvention = wasmCallingConvention().callInformationFor(signature, CallRole::Caller);
@@ -167,7 +167,7 @@ JSC_DEFINE_JIT_OPERATION(operationJSToWasmEntryWrapperBuildReturnFrame, EncodedJ
 
     auto access = [sp, callFrame]<typename V>(auto* arr, int i) -> V* {
         dataLogLnIf(WasmOperationsInternal::verbose, "fp[", (&reinterpret_cast<uint8_t*>(arr)[i / sizeof(uint8_t)] - reinterpret_cast<uint8_t*>(callFrame)), "] sp[", (&reinterpret_cast<uint8_t*>(arr)[i / sizeof(uint8_t)] - reinterpret_cast<uint8_t*>(sp)), "](", reinterpret_cast<V*>(arr)[i / sizeof(V)], ")");
-        return &reinterpret_cast<V*>(arr)[i / sizeof(V)];
+        return &reinterpret_cast<V*>(arr)[i / static_cast<int>(sizeof(V))];
     };
 
     if (signature.returnsVoid())
@@ -300,7 +300,7 @@ JSC_DEFINE_NOEXCEPT_JIT_OPERATION(operationGetWasmCalleeStackSize, UCPUStrictInt
 JSC_DEFINE_JIT_OPERATION(operationWasmToJSExitMarshalArguments, void, (void* sp, CallFrame* callFrame, void* argumentRegisters, JSWebAssemblyInstance* instance))
 {
     auto access = []<typename V>(auto* arr, int i) -> V* {
-        return &reinterpret_cast<V*>(arr)[i / sizeof(V)];
+        return &reinterpret_cast<V*>(arr)[i / static_cast<int>(sizeof(V))];
     };
 
     // We need to set up them immediately before potentially throwing anything.
@@ -443,7 +443,7 @@ ALWAYS_INLINE void assertCalleeIsReferenced(CallFrame* frame, JSWebAssemblyInsta
 JSC_DEFINE_JIT_OPERATION(operationWasmToJSExitMarshalReturnValues, void, (void* sp, CallFrame* callFrame, JSWebAssemblyInstance* instance))
 {
     auto access = []<typename V>(auto* arr, int i) -> V* {
-        return &reinterpret_cast<V*>(arr)[i / sizeof(V)];
+        return &reinterpret_cast<V*>(arr)[i / static_cast<int>(sizeof(V))];
     };
 
     void* registerSpace = sp;
